+++ restcomm/restcomm.core/src/test/java/org/restcomm/connect/core/service/utils/UriUtilsTest.java
@@ -1,0 +1,35 @@
+/*
+ *  TeleStax, Open Source Cloud Communications
+ *  Copyright 2011-2018, Telestax Inc and individual contributors
+ *  by the @authors tag.
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  under the terms of the GNU Affero General Public License as
+ *  published by the Free Software Foundation; either version 3 of
+ *  the License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>
+ */
+package org.restcomm.connect.core.service.utils;
+
+import java.net.URI;
+import org.junit.Test;
+import org.restcomm.connect.core.service.util.UriUtils;
+
+public class UriUtilsTest {
+
+    public UriUtilsTest() {
+    }
+
+    @Test(expected = RuntimeException.class)
+    public void testNoConnector() throws Exception {
+        UriUtils.resolve(new URI(""));
+    }
+
+}
